Behavior Tree in State Machine. Setup for RoboPadron
First draft of behavior tree infrastructure.

+++ Assets/AivoBehaviorTree/ActionNode.cs
@@ -4,14 +4,14 @@ namespace AivoTree
 {
     public class ActionNode<T> : TreeNode<T>
     {
-        private readonly Func<long, T, AivoTreeStatus> _fn;
+        private readonly Func<float, T, AivoTreeStatus> _fn;
 
-        public ActionNode(Func<long, T, AivoTreeStatus> fn)
+        public ActionNode(Func<float, T, AivoTreeStatus> fn)
         {
             _fn = fn;
         }
         
-        public AivoTreeStatus Tick(long timeTick, T context)
+        public AivoTreeStatus Tick(float timeTick, T context)
         {
             return _fn(timeTick, context);
         }
